requirements.txt
groq 
python-dotenv
SpeechRecognition 
pydub 
pyaudio 
gTTS 
gradio     
elevenlabs 
streamlit
